+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### [13.15.0](https://kaos.sh/ek/13.15.0)
+
+- `[color]` Added `fmt.GoStringer` support for structs
+- `[req]` Added more authentication methods (`OAuth`, `Digest`, `VAPID`, `AWS4`, `API Key`)
+- `[req]` Added support of custom authentication methods
+- `[color]` Code refactoring
+
 ### [13.14.2](https://kaos.sh/ek/13.14.2)
 
 - `[log]` Added variable `LogLevels` with all supported log level names